feat(core/objpool): add objpool_alloc_with_id and objpool_get_by_id
This commit introduces two new methods to the objpool API.
The first method, objpool_alloc_with_id, enables the allocation of an
object within a pool and returns the index where the object was allocated.
The second method, objpool_get_by_id, allows retrieving a previously
allocated object using the index of the pool where it was initially allocated.

Signed-off-by: João Peixoto <joaopeixotooficial@gmail.com>