fix(core/vm): remove unnecessary cpu_idle and vcpu_arch_run calls
This commit is structured into two parts:
- Removed unnecessary cpu_idle call from vcpu_run to prevent
redundant invocations. This change eliminates continuous calls to cpu_idle in
scenarios where the profile implementation does not jump to a predefined
wake-up point but instead returns directly.
In such cases, the stack was manually rewound, and control was transferred
to the cpu_idle_wakeup function, which then re-entered vcpu_run, leading to
repeated invocations of cpu_idle.
- Removed the vcpu_arch_run call to avoid repeating the same logic for ARM and
RISCV architectures. Instead now we call directly the vcpu_arch_entry if the
core is ready to run.

Signed-off-by: João Peixoto <joaopeixotooficial@gmail.com>