fix(core/remio): use of zero-sized array
The specific error is triggered in the macro expansion for
OBJPOOL_ALLOC when the value CONFIG_REMIO_DEV_NUM assumes the
value of 0.
To solve this problem, we allocate only one slot if the user
does not configure any Remote I/O device.

Note: While we could use the #if macro to define the object pool
only when a Remote I/O is configured, this approach would clutter
the code with constant checks for device configuration whenever
accessing certain variables.

Signed-off-by: João Peixoto <joaopeixotooficial@gmail.com>